Rename jogl2 to jogl.
Rename nifty-renderer-jogl2 to nifty-renderer-jogl. Rename the actual
module directory and its reference in the main parent pom.

Why have version numbers in the module name? It makes no sense. Not only
would it be tedious to maintain as the version changed, but it's
inconsistent with all the other rendering modules. The whole point of
each module having it's own pom file is to specify the current version
in there.