element.show() not propagates EffectEventId.onShow to children anymore
When you show() an element in the past the EffectEventId.onShow was
triggered for all children, even for invisible ones. So we'd end up with
invisible child elements with an active onShow effect which would
prevent an actual show() call on that child elements.

This commit will now not propagate EffectEventId.onShow to the children
anymore.