chore: added pathBegin() and pathEnd() methods to RenderDevice
and enabled two pass rendering for lines and circles in preparation of
having paths that consist of line as well as circle parts. In the first
pass only the alpha value is calculated and stored in an alpha texture.
In the second pass a quad is rendered using the line color and taking
the alpha texture into account.