chore: move NiftyLineParameters and NiftyArcParameters closer to SPI
These classes primarly exists to communicate the line/arc parameters to
the NiftyRenderDevice. On the other hand they are used in the internal
package to store these parameters. But they don't belong into the api
package since user code does not use these classes (and should not know
about them). Therefore they have now been moved to a subpackage in the
spi package.

I'm still not too happy with that but at the moment it's a better place
than the api package :)