fix: Really fix hint effect completely this time
We still have to initially hide the hint. This was missing from the last
commit which lead to the possibility that a hint was initially been
displayed. This occured in the controls example demo where the hint was
suddenly displayed in the top left corner when you hovered over the menu
buttons on the top.

Now we initially hide the hintLayer.

#246