fix: element.show() needs to trigger onShow events for visible elements
The commit 77059d2544d3825aaa680491683b377aa1b1fb41 introduces an issue
with the controls example. All the main panels (and the hints too)
didn't fade in properly because of the commit mentioned above.

The show() method has now been changed back to the original code that
will use startEffect() instead of startEffectWithoutChildren(). Doing
that means we'll - again - iterate through all child elements.

But this time we'll not start the onShow effect for elements that are
invisible. So we fix the original problem where it was possible that
invisible child elements with an active onShow effect prevented an
actual show() call and still keep the remaining functionaility intact -
which fixes the controls example.