feat: Add initial version of Nifty CSS support
This commit adds basic CSS support to Nifty. You can apply a CSS file
to the main Nifty instance using the new method applyCSS():

nifty.applyCSS(<inputStream>);

To make it very easy to add CSS properties to any class and to custom
controls later this commit adds the @NiftyCssProperty annotation. You
simply add a getter and setter method for your property to your class
and annotate either the getter or setter with the @NiftyCssProperty
annotation to define the name of the property in the css file as well as
a NiftyCssStringConverter instance if the property you make available as
a CSS property is not a String.

Example:

@NiftyCssProperty(name = "background-color", converter =
NiftyCssStringConverterNiftyColor.class)
public void setBackgroundColor(final NiftyColor color)

Additionally there is a new setCssClass() setter available at the
NiftyNode to set the css class:
  
NiftyNode niftyNode = ...;
niftyNode.setCssClass("my");

For parsing CSS files we use Phil Browns "jCSS-Parser":
https://github.com/phil-brown/jCSS-Parser

And for processing CSS-selectors we use Christer Sandbergs
"css-selectors" project:
https://github.com/chrsan/css-selectors
https://github.com/chrsan/css-selectors/pull/13

Both have been included directly as source because jCSS-Parser is not
available in the Maven repository and css-selectors is available but not
yet in the version we require (s.a. the pending pull request linked
above).