Renamed "CSS" to "Style" as discussed with void256
This renames the CSS related part of the system to be known as "styles". Since its not a proper CSS file as a browser would use it, it's a nifty-specific CSS system.

The original files regarding default CSS handlers and systems have been left untouched (like the parsers and external libraries).

Javadocs have been changed accordingly as well.