feat(css): add proper getNodesByTagName() support and support write-only
getNodesByTagName() now uses the SimpleName of the class as the tag and
correctly uses it to compare the TagName requested.

Additionally @NiftyStyleProperty now supports read-only properties as
well as write-only properties. It's not not required to have both, a
getter and a setter defined. Having a getter, a setter or both is now
allowed. Just make sure to annotate it correctly. In case you have both,
a setter and a getter annotate either one is enough.