feat: added CSS style resolve for controls and example for Label control
Controls (implmentations of NiftyControl) can now be styled as well. The
NiftyNode a Control has been applied to now keeps a reference of the
NiftyControl for that and when a NiftyNode is styled it's referenced
NiftyControl is styled as well.

This commit adds @NiftyStyleProperty to the Label control. Another CSS
example has been added for the Label control as well where all
attributes of the Label are set via CSS.

Besides a default constructor NiftyStyleStringConverter implementations
now can provide a constructor that takes a Nifty instance as it's
parameter. This was necessary for the NiftyStyleStringConverterNiftyFont
which needs access to the Nifty instance to create Fonts.