feat: add setBackgroundGradient to NiftyNode incl. CSS3 support
This commit adds a setBackgroundGradient() method to the NiftyNode and
the NiftyCanvasPainterDefault class. Additionally it's now possible to
specifiy the "background-gradient" CSS property to style this from CSS.
This is actually parsed according to CSS3 so you can use something like:

background-gradient: linear-gradient(to top left, #737373, #343434 37%,
#000000 40%, #000000 44%, red);

The NiftyColor class now supports all of the CSS3 color constants. The
constants that the NiftyColor class provided have now been changed to
lowercase, so instead of NiftyColor.RED() you now have to use
NiftyColor.red() and so on.

The NiftyRenderDevice method renderLinearGradientQuads() now expects
simple parameters instead of the NiftyLinearGradient class.

Additionally this commit adds a preliminary Button control but this is
still pretty much work in progress. Stay tuned.