fix: Button example not working when showing FPS
The FPS display uses internally NiftyNodes to display its data. For this
purpose another rootNode is created with a label for the FPS counter.
(Note: Nifty supports multiple root nodes without any issues).

What happens in the example was that the ":root" style was applied to
all root nodes including the FPS overlay.

This commit adds a niftyPrivateNode attribute to the InternalNiftyNode
and the FPS display will mark the root node as private. niftyPrivateNode
will not be displayed in the regular getStateInfo output and the root
node will not be returned as root node when being requested by CSS
":root" selector.

fixes #305