feat: add support for directly supplying a ResourceBundle again
Additionally to the existing method another addResourceBundle() method
has been added that takes a ResourceBundle instance.

Please note a few things:

- You can't mix using the regular baseName based method and the
ResourceBundle based method for the same id. The last call will win when
using the same id.
- You can and probably should call the new addResourceBundle() method
with different instances of ResourceBundle that have been initialised
with the Locales you want to support. It's okay to only use a single one
too but in that case you can't switch the locale with nifty.setLocale()
of course.
- The first call to the new addResourceBundle() method will be used when
nifty.setLocale() is called with null - the default locale basically.

#303