fix: remove NPE when processing CSS with node hierachies with no root
Some internal node trees like the one used to display the FPS counter
don't have a public root node and appear from the view of the CSS
selector processor as node hierachies without a root node. This commit
will fix the NPE that was logged when such cases were processed.