feat: make NiftyNode state aware to support different values for Hover
A NiftyNode will always have a NiftyNodeState.Regular. All properties
you set on a NiftyNode will be a part of NiftyNodeState.Regular.
Currently all properties you can directly apply with CSS will set this
state as well.

So if you call f.i. setBackgroundColor() then this background color will
be the regular background color that will be used when the node is
displayed.

Additionally a new setBackgroundColor() method has been added with an
additional NiftyNodeState parameter that you can use to set a background
color for the NiftyNode in NiftyNodeState.Hover state. So as soon as the
NiftyNode will have a hovering pointer the background color associated
with the Hover state will be used.

In the future we will be add additional states like Active (or Armed)
and Focus and we plan to support setting these with the CSS pseudo
classes that exist for that purpose like ":hover", ":focus", ":active"
(not yet supported).

Additionally this commit adds a forceState() method to NiftyNode.
Usually Nifty will internally make sure that the properties for the
correct state will be used. However sometimes you might want to lock a
certain state, f.i. "Hover". This is mainly a debugging or a feature for
an editor where you want to view a NiftyNode in a certain state instead
of always hovering the Node just to test your new style. If you call
forceState() then Nifty will not change that state anymore, e.g. Hover
will not be triggered anymore - only the forcedState is being used.