fix(Java2D): Initialize graphics and graphics2dHelper in constructor
Both were initialized lazily in the beginFrame() method which is only
called when rendering Nifty. Therefore it was possible that the graphics
and graphics2dHelper member variables were null before the first call of
nifty.render(). However for getting RenderFont attributes (like the
width) the Java2D implementation required the graphics instance!

But it is very likely that you call fromXml() before the very first
frame being renderedn and you'd end up with a none-initialized graphics
member variable and therefore a font.getWidth() call returns 0.

I think this lead to a lot of issues one reported was that wrap="true"
did not work unless render() was called at least once.

So this commit will init both variables directly in the constructor
fixing this!

fix #317 wrap="true" does not work in Java2D renderer