refactor: seperated tree sync and state sync into seperated classes
Syncing the NiftyNode tree with the RenderNodes is now seperated from
the syncing of the render states. In the old code everything was messed
up into a single class and was quite confusing. I think it's a bit
clearer now.

Additionally this fixes some bugs when nodes where removed from the
NiftyNode tree which was not correctly represented in the RenderNode
tree.

And nodes with zero width or height will now be correctly removed from
the RenderNode tree as well :)