fix: fix screenToLocal() and localToScreen() for grandchildren
Both methods used the parent local transformation which is not correct
(and only worked for one ancestor). When calculating screenToLocal (or
its inverse) the parents screenToLocal transformation matrix has to be
used to correctly take into account the transformations of all parents -
not only the first.

This commit fixes this issue and cleans up some of the examples as well.
Take a look at the PointerEventsScreenToLocal example which illustrates
the now correct behaviour.