feat: integrate version 2.0 of se.fishtank:css-selectors:2.0 project
The direct source inclusion of the 1.x version has been removed and the
selector adapter has been changed to the 2.0 API of the css-selectors
project.

This works so far but we only use a small subset of the selector API so
further changes might be necessary to the new InternalNiftyNodeDOMNode
adapter that interfaces the NiftyNode to the css-selector DOMNode.