feat: new NiftyTree added (work in progress)
This is a first commit at establishing a new way to organize NiftyNodes.

In the long run we're trying to replace the existing "hero" NiftyNode class
with smaller NiftyNode interface implementations. The first goal in reducing
the work done in the current NiftyNode will be to extract the tree hierachy
management from NiftyNode. In the future this task will be performed by the
NiftyTree class (work in progress name - need to find a better one ;)

The NiftyTree provides methods to build and manage a tree structure of
NiftyNodes. Please note that for the time being a new NiftyNode interface
has been added in the node package. This is different from the existing
NiftyNode class. Once we have the full NiftyTree in place the first step will
be to remove all child management from NiftyNode and add the NiftyTree
to the public Nifty API. Stay tuned :)