feat: static Nifty setting for default NiftyInputMapping.
ElementType will now use a default class type accessed by name via a static property specified in NiftyDefaults.

Unfortunately, there is no way to inject the setting into ElementType at this time due to no instance of Nifty (or Screen, ScreenController, etc) being accessible. This is a "maximum compatibility" approach, but it should be noted that it is most certainly non-optimal. It would be better if the change could be per each Nifty instance.

Regardless, this should allow users to override Nifty's default input mapper, so that mappings can be applied globally.

Note: the danger here is that the setting will persist throughout the JVM rather than just a single Nifty or Screen instance. Again, not optimal.