chore: store NiftyNodeImpl in NiftyTree (experimental)
NiftyTreeNode now directly stores NiftyNodeImpl<? extends NiftyNode>
instances.

NiftyNodeImpl is the implementation part of a NiftyNode and will get
a reference to the actual API NiftyNode interface.

Custom NiftyNode and NiftyNodeImpl will need to be registered with
Nifty. Build-In NiftyNodes will automatically registered.

When NiftyNodes will be added to Nifty, Nifty will link the NiftyNode
added and the NiftyNodeImpl using the registered mapping.

InternalNiftyTree has been extended with a lot of methods to access
NiftyNode and NiftyNodeImpl directly.