Fix Slick mouse cursor bug.
The bug:

- SlickBatchRenderBackendFactory is creating a BatchRenderBackendInternal,
which simply calls MouseCursor#enable in
BatchRenderBackendInternal#enableMouseCursor for custom mouse cursors.
This works for every backend except for Slick because MouseCursor must
be cast to a SlickMouseCursor, and then
SlickMouseCursor#enable(GameContainer container) must be called, NOT
MouseCursor#enable, which throws an UnsupportedOperationException after
extensive refactoring of the custom mouse cursor system in Nifty (see
commit 2c4b5d0ac6ccad64b9636a2b8ef3283e27ffc5f0 for details).

The fix:

- Get rid of SlickMouseCursor#enable(GameContainer container) &
SlickMouseCursor#disable(GameContainer container), as they violate the
encapsulation of the newer MouseCursor#enable & MouseCursor#disable
methods, respectively.

- Remove AbstractNativeSlickMouseCursor#enable &
AbstractNativeSlickMouseCursor#disable implementations, as they only
throw UnsupportedOperationException's (due to a leaky abstraction).

- Add LwjglCursorSlickMouseCursor#enable &
LwjglCursorSlickMouseCursor#disable methods, that implement
MouseCursor#enable & MouseCursor#disable, respectively. Delete
LwjglCursorSlickMouseCursor#enable(GameContainer container) &
LwjglCursorSlickMouseCursor#disable(GameContainer container), as they
are removed from the implemented SlickMouseCursor interface.

- Require a GameContainer to be passed into the
LwjglCursorSlickMouseCursor constructor, that can be used to enable &
disable the mouse cursor via #enable & #disable, respectively. This
removes the need for #enable(GameContainer container) &
can be treated as a MouseCursor instead of a SlickMouseCursor when
enabling & disabling, just like every other backend's mouse cursors
(patching the leaky abstraction).

- Create SlickBatchRenderBackend, which delegates every single method to a
BatchRenderBackendInternal, EXCEPT for the rendering of non-native
SlickMouseCursor's. Require a GameContainer in the
SlickBatchRenderBackend constructor, as it must be passed to the
SlickMouseCursor#render in #endFrame. This is the still-leaky part of
the MouseCursor abstraction, as MouseCursor does not have a #render
method, but non-native SlickMouseCursor's must have #render called every
frame, so cast MouseCursor to SlickMouseCursor (if possible) and render
it in #endFrame.

- Create a SlickBatchRenderBackend in SlickBatchRenderBackendFactory
instead of a BatchRenderBackendInternal, to get special mouse rendering
support for SlickMouseCursor's.

- Require a GameContainer in SlickBatchRenderBackendFactory#create, which
gets passed to SlickBatchRenderBackend & SlickMouseCursorFactory.

- Update all internal uses of SlickBatchRenderBackendFactory#create by
passing in the GameContainer.

- Require SlickMouseCursorLoader#loadCursor &
LwjglCursorSlickMouseCursorLoader#loadCursor to take a GameContainer, as
it must be passed to the returned LwjglCursorSlickMouseCursor, so that
it can be enabled & disabled.

- Require SlickMouseCursorLoaders#loadCursor to take a GameContainer, as
it must be passed to the returned LwjglCursorSlickMouseCursor, so that
it can be enabled & disabled.

- Fix SlickRenderDevice (deprecated) mouse cursor support to work with
the aforementioned changes.

API Changes:

- All users of the Slick batch rendering backend must pass a
GameContainer to SlickBatchRenderBackendFactory#create when
instantiating Nifty, i.e.:
  new Nifty(new BatchRenderDevice(SlickBatchRenderBackendFactory.create(gameContainer)), ...);

- If users of the Slick backend uses a convenience class such as
NiftyGameState or NiftyOverlayGameState, this will be done
automatically.

GitHub Issue References:

Fixes #324
Fixes #355