feat: Added node accessor loading using hacky service loader
This version uses a slightly altered version of the java.util.ServiceLoader
that allows instantiating classes that are not public. This allows loading
the accessors directly without the need of any additional service class.
The API is really clean this way, how ever the service loader is a little
hacky because it ignores the access level of the accessor classes.

The tests are adapted and work with those changes.