refactor: add "Nifty" prefix to Point, Rect and Size classes
Just to be consistent with all the other Nifty* classes and to make sure
we don't conflict with existing Point classes (like java.awt.Point) even
though they are in a different package. This way everything in the Nifty
universe starts with Nifty - even if it is a bit more to type.