feat: new internal renderer added
This is still heavy work in progress but it seems to already render the testcases ;) There is still a lot room left for optimisations but the basic algorithm is now in place.

The main idea is to separate the screen into individual smaller textures (currently 256x256) which we call buckets. The idea is to later render only these buckets that have changed. To do that we store a list of RenderNodes into each bucket when the RenderNode rectangle intersects the bucket rectange. This way we have a list of buckets and each of the buckets contains a list of all the RenderNodes that overlap this bucket.

A RenderNode is a representation of a NiftyContentNode with resolved/transformed coordinates that is backed by a NiftyCanvas texture.

So in the end the whole NiftyTree of Nodes is resolved down to a view RenderNodes and these are placed into the buckets where they intersect. A RenderNode can be stored in multiple buckets when it overlaps several of them.

The idea is to later track all changes to RenderNodes and to know per bucket if a list of render nodes (or the RenderNodes themself) have changed and when that occurs we need to re-render the bucket with the changed content. Since each bucket is stored in a texture the idea is that local changes will only require the re-rendering of a view buckets while we can keep other buckets the same and simply re-use their stored texture data as is.

When nothing has changed we can get away with rendering a view textures :)