fix: keep original renderOrder of elements
When RenderBucketRenderNodes were added/removed dynamically
from RenderBuckets it was possible that their original order changed.
This could lead to rendering errors in the order of elements.

This has now been fixed by sorting the RenderBucketRenderNodes
ascending in their original order as they appear in the NiftyTree.

Note: We should later allow the user to influence the new renderOrder
attribute we've added to the RenderBucketRenderNode. But for now we'll
keep the original order.