fix: NiftyTree - Fixed implementation of Iterable
The old implementation of the iterable returned by child nodes did not
ensure that the iteration starts from the beginning in case a new
iterator is requested as it is expected from a non-consuming iterator.

The new implementation creates a new iterator instance each time the
iteration is requested to fulfill this requirement. Currently this
problem does not occur because each iterable instance is only
iterated once. But this may cause trouble in future.