chore: methods added to set matrix value on existing instance
setTranslate, setScale and setRotate added to modify a Mat4
instance in-place. These methods should be used carefully
since they expect the matrix to already be either an identity
matrix or a translate, scale or rotation matrix.

If you apply a setTranslate to a scale matrix you'll probably
not get the expected matrix :)

Why add these methods then anyway? The new methods should
be used to prevent uneccessary garbage to be generated since
existing Mat4 instances can be reused easily this way.