fix: transformations are now applied correctly and are simplified as well
The NiftyTransformationNode and the NiftyContentNode now keep
track of the current LayoutRect using the NiftyState-mechanism. This
way we can calculate the relative layout position of these nodes. Relative
to the parent Transformation- or ContentNode.

This allows us now to again store the complete transformation matrix
in the NiftyTransformationNode which simplifies the way in which we
calculate the final transformation matrix for a NiftyNode!

Additionally this fixes a problem in the old way where transformations
were incorrectly applied to individual Nodes when instead they should
have been applied to the group of Nodes.

This commit introduces UseCase_a07_GroupedTransformation exactly
to demonstrate this now corrected behaviour.

Kudos to mkaring for figuring out what the problem was and helping
to fix it :)