feature: make NiftyConfiguration public and cleanup Nifty class
Configuration parameters have now been extracted into a NiftyConfiguration
class that is now an optional parameter when Nifty is instantiated.

For the time being the current configuration values can only be set
once (when Nifty is instantiated) and cannot be changed later.

UseCaseRunner in the nifty-examples project has now an additional
run() method to use a custom NiftyConfiguration.

NiftyStatisticsMode has been removed for now.