fix: MouseOverHandler processing for on hover effects
In some former commit hover events were changed so that they were reported
as being processed events (just like regular click events).

Originally the MouseOverHandler stopped processing completely when such
"processed" events were detected while processing elements "under" the mouse
cursor.

However, because of that hover effects were marked internally as being used
but were never reset. The next time they were supposed to be activated the
effect start method was never called because internally the effect was still
flagged as active.

This commit changes two things:

1. Processing will not stop at the first "processed" mouse over event
completely. Only the processing of mouse over events will be skipped in this
case. The regular processing of the event for not mouse over elements will
continue.

2. Originally the mouse event was forwarded to the list of mouse over elements
first and then to the list of elements that can process mouse events but
currently were not "under" the mouse cursor. The latter step was necessary so
that eventually attached hover effects could be internally correctly reset.
However, executing the mouse over elements first and the regular none mouse
over elements after that would lead to execution of the onStartHover effects
of new elements before the onEndHover effects of the old ones. Since the
effect-processing checks in both cases if the mouse is actually inside the
elements again it should be save to switch the both around. So now, the event
is first forwarded to the elements that are not part of the mouse over list and
then after that to the elements that are part of the mouse over element list. This
way the onStart/onEndHover elements are executed in the expected order.

However, there is a slight chance that this change has other unwanted side-
effects. I'm not aware of any right now though.

#412 OnStartHover and OnEndHover are not working properly when moving the mouse fast between hoverable elements