refactor: fit PathRenderer internals more closely to HTML5 canvas spec
mainly this introduces the SubPath inner class in PathRenderer so that
the checks of the HTML5 canvas spec are more closely resembled

additionally the pathArcs() method has been removed from the
NiftyRenderDevice (and all of its implementations) since currently
we render arcs using lines and pathArcs() is not being used anymore.
This will make it easier for new NiftyRenderDevice implementations.