chore: directly integrate org.bushe.eventbus-1.4 source
The original project appears to be not updated anymore. In issue #364
a problem in that lib has been raised that will need to be fixed.

Since we don't expect the original project to be updated anymore (or
at least not in the foreseeable future) we choose the brute force solution:
include the source of the eventbus project directly and then later fix
the issue with it in our codebase.

This commit is the first step: include the source into our codebase.

 #364: Application not closing cleanly unless I delete some elements