chore: first steps for improving font API and implementation (wip)
In the future font rendering is supposed to happen again in NiftyRenderDevice
implementations. This will allow us a greater flexibility in supporting different
kinds of font rendering technologies compared to the original plan to use
bitmap fonts only.

As a first step the existing NiftyFont class has been changed to an interface
and is now part of the SPI.

Please note: Currently the NiftyFont still contains methods to perform the
rendering and request informations about the font. This is very likely
changing in the future and is only left in by now so that we don't break
existing font rendering. However, the idea is that in the future NiftyFont is
just a representation (a handle) to the actual underlying font resource and
all the rendering will be done in a NiftyRenderDevice impl (which created the
NiftyFont instance and knows how to render it).

Additionally the org.jglfont.jglfont-core dependency has been removed from
Nifty and is now included in source directly. It is now a part of the nifty-
renderer-opengl project and for now the only kind of font it is supporting.

More changes towards font rendering will happen in the upcoming days.
Stay tuned ;)