Disable double click in slider control
Double clicking on a slider just swallows the click because it is a multiclick event now and the slider is only listening for single clicks. This should make clicks, multiclicks and click holding do the same. I don't know if there are other controls that would benefit of something like this, but in my case, this was driving me crazy.